feat(#93): move status module into mvnw module
